+++ backend/model/model.go
@@ -856,7 +856,6 @@ type Field struct {
 	// 'email@email.com'
 	Value     string    `gorm:"uniqueIndex:idx_fields_type_name_value_project_id"`
 	ProjectID int       `json:"project_id" gorm:"uniqueIndex:idx_fields_type_name_value_project_id"`
-	Sessions  []Session `gorm:"many2many:session_fields;"`
 	Timestamp time.Time `json:"timestamp"`
 }