+++ backend/private-graph/graph/auth_client.go
@@ -47,7 +47,7 @@ type FirebaseAuthClient struct {
 	authClient *auth.Client
 }
 type OAuthClient struct {
-	domain       string
+	domains      []string
 	oidcProvider *oidc.Provider
 	oauthConfig  *oauth2.Config
 }
@@ -474,7 +474,7 @@ func (c *OAuthAuthClient) updateContextWithAuthenticatedUser(ctx context.Context
 	prov, err := c.getOIDCProvider(req)
 	if err != nil {
 		log.WithContext(req.Context()).WithError(err).Error("error getting oidc provider")
-		http.Error(w, err.Error(), http.StatusBadRequest)
+		return ctx, e.New("error getting oidc provider")
 	}
 
 	verifier := prov.Verifier(&oidc.Config{ClientID: clientID})
@@ -482,7 +482,7 @@ func (c *OAuthAuthClient) updateContextWithAuthenticatedUser(ctx context.Context
 	if err != nil {
 		log.WithContext(ctx).WithField("token", token).WithError(err).Info("invalid user token")
 		c.handleLogout(w, req)
-		return ctx, nil
+		return ctx, e.New("invalid user token")
 	}
 
 	// Extract claims
@@ -490,7 +490,7 @@ func (c *OAuthAuthClient) updateContextWithAuthenticatedUser(ctx context.Context
 	if err := idToken.Claims(&claims); err != nil {
 		log.WithContext(ctx).WithField("token", token).WithError(err).Info("invalid user claim")
 		c.handleLogout(w, req)
-		return ctx, nil
+		return ctx, e.New("invalid user claim")
 	}
 
 	// check that the oidc email domain matches allowed domains
@@ -502,9 +502,9 @@ func (c *OAuthAuthClient) updateContextWithAuthenticatedUser(ctx context.Context
 	domain := parts[1]
 
 	var validated bool
-	var domains = lo.Map(lo.Values(c.oauthClients), func(item *OAuthClient, _ int) string {
-		return item.domain
-	})
+	var domains = lo.Flatten(lo.Map(lo.Values(c.oauthClients), func(item *OAuthClient, _ int) []string {
+		return item.domains
+	}))
 	for _, dom := range domains {
 		if dom == domain {
 			validated = true
@@ -644,10 +644,14 @@ func NewOAuthClient(ctx context.Context, store *store.Store) (*OAuthAuthClient, 
 			Scopes: []string{oidc.ScopeOpenID, "profile", "email"},
 		}
 
-		oauthClients[ssoClient.ClientID] = &OAuthClient{
-			domain:       ssoClient.Domain,
-			oidcProvider: provider,
-			oauthConfig:  &oauth2Config,
+		if _, ok := oauthClients[ssoClient.ClientID]; ok {
+			oauthClients[ssoClient.ClientID].domains = append(oauthClients[ssoClient.ClientID].domains, ssoClient.Domain)
+		} else {
+			oauthClients[ssoClient.ClientID] = &OAuthClient{
+				domains:      []string{ssoClient.Domain},
+				oidcProvider: provider,
+				oauthConfig:  &oauth2Config,
+			}
 		}
 	}