+++ backend/worker/worker.go
@@ -305,6 +305,8 @@ func (w *Worker) processPublicWorkerMessage(ctx context.Context, task *kafkaqueu
 	span, ctx := util.StartSpanFromContext(ctx, "worker.processPublicWorkerMessage")
 	defer span.Finish()
 
+	ctx, cancel := context.WithCancel(ctx)
+
 	done := make(chan struct{})
 	defer func() {
 		close(done)
@@ -322,6 +324,7 @@ func (w *Worker) processPublicWorkerMessage(ctx context.Context, task *kafkaqueu
 		ticker := time.NewTicker(1 * time.Second)
 		defer ticker.Stop()
 
+		canceled := false
 		for {
 			select {
 			case <-done:
@@ -335,6 +338,11 @@ func (w *Worker) processPublicWorkerMessage(ctx context.Context, task *kafkaqueu
 					WithField("partition", task.KafkaMessage.Partition).
 					WithField("duration", duration).
 					Warnf("Long running task")
+				if !canceled && duration > 10*time.Minute {
+					log.WithContext(ctx).Errorf("Cancelling long running task")
+					cancel()
+					canceled = true
+				}
 			}
 		}
 	}()
@@ -847,10 +855,17 @@ func (w *Worker) processSession(ctx context.Context, s *model.Session) error {
 		return w.excludeSession(ctx, s, backend.SessionExcludedReasonNoActivity)
 	}
 
-	visitFields := []model.Field{}
-	if err := w.Resolver.DB.WithContext(ctx).Model(&model.Session{Model: model.Model{ID: s.ID}}).Where("Name = ?", "visited-url").Where("session_fields.id IS NOT NULL").Order("session_fields.id asc").Association("Fields").Find(&visitFields); err != nil {
-		return e.Wrap(err, "error querying session fields for determining landing/exit pages")
+	fields, err := w.Resolver.Redis.GetSessionFields(ctx, s.SecureID)
+	if err != nil {
+		return err
 	}
+
+	visitFields := lo.Filter(fields, func(field *model.Field, _ int) bool {
+		return field.Name == "visited-url"
+	})
+	sort.Slice(visitFields, func(i, j int) bool {
+		return visitFields[i].Timestamp.Before(visitFields[j].Timestamp)
+	})
 
 	pagesVisited := len(visitFields)
 
@@ -890,7 +905,7 @@ func (w *Worker) processSession(ctx context.Context, s *model.Session) error {
 			{Key: "exit_page", Value: exitPage.Value, Timestamp: exitPage.CreatedAt},
 		}
 
-		if err := w.PublicResolver.AppendProperties(ctx, s.ID, sessionProperties, pubgraph.PropertyType.SESSION); err != nil {
+		if err := w.PublicResolver.AppendProperties(ctx, s.SecureID, sessionProperties, pubgraph.PropertyType.SESSION); err != nil {
 			log.WithContext(ctx).Error(e.Wrapf(err, "[processSession] error appending properties for session %d", s.ID))
 		}